backend/requirements.txt
fastapi
uvicorn
supabase
PyPDF2
python-docx
python-multipart
beautifulsoup4
httpx
python-jose[cryptography]
pydantic
python-dotenv
